% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate.R
\name{duplicate}
\alias{duplicate}
\title{Duplicate a function. Returns a carbon copy of the original, so that original
can be modified safely.}
\usage{
duplicate(x)
}
\description{
Duplicate a function. Returns a carbon copy of the original, so that original
can be modified safely.
}
\examples{
\dontrun{
   fn <- function(x) cat(x, "\\n")
   fn2 <- duplicate(fn)
   fn <- function(x) cat(x, "woof","\\n")
   fn("hello") # hellowoof
   fn2("hello") # hello
}
}

