% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_called.R
\name{expect_called}
\alias{expect_called}
\title{Expect that a function was called while executing an expression.}
\usage{
expect_called(fn, expr, package_name, was_called = TRUE)
}
\arguments{
\item{fn}{function. The function that was expected to get called. If not
specified with \code{::}, package_name param must be specified.}

\item{expr}{expression. The expression to execute.}

\item{package_name}{string. The package where fn lives. If not specified, fn must
be package_name::fn.}

\item{was_called}{logical. If \code{FALSE}, negate the meaning: \code{fn}
should \emph{not} have been called. By default \code{TRUE}.}
}
\value{
invisible \code{TRUE}.
}
\description{
Expect that a function was called while executing an expression.
}
\examples{
\dontrun{
  expect_called(base::print, print("Hello")) # Will pass
  expect_called(print, print("Hello"), "base") # Will pass
  expect_called(base::print, cat("Hello")) # Will fail
  expect_called(base::print, cat("Hello"), was_called = FALSE) # Will pass
}
}

