% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.directory}
\alias{is.directory}
\title{Check whether a path is a directory.}
\usage{
is.directory(path)
}
\arguments{
\item{path}{character. The name of the path.}
}
\description{
Check whether a path is a directory.
}

