% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_httr_response.R
\name{mock_httr_response}
\alias{mock_httr_response}
\title{Mocks an httr response}
\usage{
mock_httr_response(status_code, content, type)
}
\arguments{
\item{content}{list.}

\item{type}{character. json or text}

\item{status}{code integer.}
}
\value{
a response object.
}
\description{
Mocks an httr response
}
\examples{
mock_httr_response(200L, list(data = "it works!"), "json")
}

