% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretend_now_is.R
\name{parse_time.character}
\alias{parse_time.character}
\title{S3 method for parse_time with \code{character} argument.}
\usage{
\method{parse_time}{character}(time)
}
\arguments{
\item{time}{is class \code{character}}
}
\description{
S3 method for parse_time with \code{character} argument.
}
