% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretend_now_is.R
\name{parse_time.Date}
\alias{parse_time.Date}
\title{S3 method for parse_time with \code{Date} argument.}
\usage{
\method{parse_time}{Date}(time)
}
\arguments{
\item{time}{is class \code{Date}}
}
\description{
S3 method for parse_time with \code{Date} argument.
}
