% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretend_now_is.R
\name{parse_time.default}
\alias{parse_time.default}
\title{Default parse_time; throws error}
\usage{
\method{parse_time}{default}(time)
}
\arguments{
\item{time}{can be any of \code{POSIXct}, \code{Date}, or \code{character}.}
}
\description{
Default parse_time; throws error
}
