% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretend_now_is.R
\name{parse_time.POSIXct}
\alias{parse_time.POSIXct}
\title{S3 method for parse_time with \code{POSIXct} argument.}
\usage{
\method{parse_time}{POSIXct}(time)
}
\arguments{
\item{time}{is class \code{POSIXct}}
}
\description{
S3 method for parse_time with \code{POSIXct} argument.
}
