% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretend_now_is.R
\name{parse_time}
\alias{parse_time}
\alias{parse_time.default}
\title{Generic parse_time}
\usage{
parse_time(time)

\method{parse_time}{default}(time)
}
\arguments{
\item{time}{can be any of \code{POSIXct}, \code{Date}, or \code{character}.}

\item{time}{can be any of \code{POSIXct}, \code{Date}, or \code{character}.}
}
\description{
Generic parse_time

Default parse_time; throws error
}
