% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pretend_now_is.R
\name{pretend_now_is}
\alias{pretend_now_is}
\title{Pretend now is some other time.}
\usage{
pretend_now_is(time, expr)
}
\arguments{
\item{time}{Date or POSIXct or character. For example,
\code{Sys.time() + as.difftime(1, units = "days")} or
\code{"1 day from now"}.}

\item{expr}{expression. The expression to evaluate while pretending
the current time is \code{time}.}
}
\value{
the value of \code{expr}.
}
\description{
This method stubs the built-in \code{\link{Sys.time}},
\code{\link{Sys.Date}}, and \code{\link{date}} methods with an overwritten
value for use in unit tests.
}
\details{
Either a \code{\link{Date}} or \code{\link{POSIXct}} object can be
provided, or an English string in the form "<number> <unit>s <from now/agO>"
such as:

\itemize{
  \item{1 hour ago}
  \item{2 months from now}
  \item{10 seconds ago}
}

These are translated into \code{Sys.time() + as.difftime(number, units = unit)}
(or \code{-} in the case of "ago").
}
\examples{
pretend_now_is(Sys.time() + as.difftime(1, units = "days"), {
 cat("It's ", Sys.time(), ", tomorrow!")
})

now <- Sys.time()
pretend_now_is("10 minutes from now", {
  stopifnot(all.equal(
   as.integer(as.difftime(Sys.time() - now, units = "minutes")),
   10))
})
}

