% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/with_options.R
\name{with_options}
\alias{with_options}
\title{Pretend Options Are Set.}
\usage{
with_options(...)
}
\arguments{
\item{...}{Named expressions are options, unnamed are evaluated.}
}
\value{
the value of the last evaluated expression.
}
\description{
Pretend certain options are set while testing an expression.
They will automatically be reset to their former values after
the expression is executed.
}
\examples{
with_options(digits = 20, print(pi)) # pi to 20 digits
print(pi) # 3.141593 -- the digits option is back to normal
}

