% Generated by roxygen2 (4.0.0): do not edit by hand
\name{assert}
\alias{assert}
\title{Explicitly state that we expect no errors to occur.}
\usage{
assert(expr, info)
}
\arguments{
\item{expr}{expression. This expression should not error.}

\item{info}{character. The error message to display.}
}
\description{
Explicitly state that we expect no errors to occur.
}

