% Generated by roxygen2 (4.0.0): do not edit by hand
\name{create_test_files}
\alias{create_test_files}
\title{Create a file structure for tests.}
\usage{
create_test_files(files)
}
\arguments{
\item{files}{list. A nested file structure. The names of the list will decide
the names of the files, and the terminal nodes should be strings which
will populate the file bodies.}
}
\value{
the directory in which this file structure exists.
}
\description{
For example, when files need to be present for a function we are testing,
it would be very cumbersome to create these files manually. Instead, we can
do the following:
}
\details{
\code{test_dir <- create_test_files(list(dir1 = list('file1',
   file2.r = 'print("Sample R code")'), file3.csv = "a,b,c\n1,2,3"))}

with the return value being a test directory containing these structured files.
Additionally, \code{create_test_files} has created an \code{on.exit} trigger
in the calling environment that will unlink everything in this test directory.
}
\examples{
\dontrun{
  test_dir <- create_test_files(list(test = 'blah', 'test2'))
  # Now test_dir is the name of a directory containing a file 'test'
  # with the string 'blah' and an empty file 'test2'.
}
}

