% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/pending.R
\name{pending}
\alias{pending}
\title{Display a yellow dot to indicate this test is pending implementation.}
\usage{
pending()
}
\description{
Note that the yellow dot allows easy use with the \code{"summary"}
reporter in the testthat package (see \code{testthat:::find_reporter}).
Using this function in a test also makes it explicitly clear that it
should be implemented shortly.
}

